api/requirements.txt
fastapi==0.87.0
uvicorn==0.20.0
python-multipart==0.0.6
opencv-python
numpy
tensorflow
pillow